Move validate_block_merkle_roots next to validate_block_body.
